+++ src/Recyclarr.Core/Compatibility/IServiceInformation.cs
@@ -3,4 +3,5 @@ namespace Recyclarr.Compatibility;
 public interface IServiceInformation
 {
     Task<Version> GetVersion(CancellationToken ct);
+    Task<string> GetAppName(CancellationToken ct);
 }