+++ src/Recyclarr.Core/Compatibility/ServiceInformation.cs
@@ -4,20 +4,30 @@ namespace Recyclarr.Compatibility;
 
 public class ServiceInformation(ISystemService api, ILogger log) : IServiceInformation
 {
-    private Version? _version;
+    private SystemServiceResult? _status;
 
     public async Task<Version> GetVersion(CancellationToken ct)
     {
-        return _version ??= await FetchVersion(ct);
+        return (await FetchStatus(ct)).Version;
     }
 
-    private async Task<Version> FetchVersion(CancellationToken ct)
+    public async Task<string> GetAppName(CancellationToken ct)
+    {
+        return (await FetchStatus(ct)).AppName;
+    }
+
+    private async Task<SystemServiceResult> FetchStatus(CancellationToken ct)
+    {
+        return _status ??= await DoFetchStatus(ct);
+    }
+
+    private async Task<SystemServiceResult> DoFetchStatus(CancellationToken ct)
     {
         try
         {
             var result = await api.GetStatus(ct);
             log.Debug("{Service} Version: {Version}", result.AppName, result.Version);
-            return result.Version;
+            return result;
         }
         catch (HttpRequestException)
         {